pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "key-extraction-aliasing-workflow"
version = "1.0.0"
description = "Key Extraction and Aliasing Workflow for Cognite Data Fusion"
authors = ["Darren Downtain <darren.downtain@cognitedata.com>"]
maintainers = ["Darren Downtain <darren.downtain@cognitedata.com>"]
readme = "README.md"
license = "MIT"
keywords = ["cognite", "cdf", "key-extraction", "aliasing", "workflow", "data-fusion"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
packages = []

[tool.poetry.dependencies]
python = "^3.12"
cognite-sdk = ">=7.83.0,<8.0.0"
pyyaml = "^6.0"
jsonschema = "^4.0.0"
aiohttp = "^3.8.0"
asyncio-throttle = "^1.0.0"
python-dotenv = "^1.0.0"
cognite-toolkit = "^0.6.80"
cognite-extractor-utils = "^7.10.0"

[tool.poetry.scripts]
key-extraction-aliasing-workflow = "main:main"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"

# Code quality
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"
ruff = "^0.1.0"
bandit = "^1.7.5"
safety = "^2.3.5"

# Pre-commit hooks
pre-commit = "^3.0.0"

# Documentation
sphinx = "^6.0.0"
sphinx-rtd-theme = "^1.2.0"
myst-parser = "^1.0.0"

# Type stubs
types-PyYAML = "*"

[tool.poetry.group.docs.dependencies]
sphinx = "^6.0.0"
sphinx-rtd-theme = "^1.2.0"
myst-parser = "^1.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"

[tool.poetry.urls]
Homepage = "https://github.com/cognitedata/key-extraction-aliasing-workflow"
Documentation = "https://key-extraction-aliasing-workflow.readthedocs.io"
Repository = "https://github.com/cognitedata/key-extraction-aliasing-workflow.git"
Issues = "https://github.com/cognitedata/key-extraction-aliasing-workflow/issues"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["key_extraction", "aliasing", "common"]
known_third_party = ["cognite", "yaml", "jsonschema", "aiohttp"]

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cognite.*",
    "yaml.*",
    "jsonschema.*",
    "aiohttp.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["modules"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Ruff configuration
[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",  # flake8-quotes
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/*" = ["ARG", "S101"]  # test files can have unused args and assert

[tool.ruff.isort]
known-first-party = ["key_extraction", "aliasing", "common"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
